+++ lib/reversetunnel/agent.go
@@ -664,21 +664,3 @@ const (
 	chanDiscoveryReq = "discovery"
 	reconnectRequest = "reconnect@goteleport.com"
 )
-
-const (
-	// LocalNode is a special non-resolvable address that indicates the request
-	// wants to connect to a dialed back node.
-	LocalNode = "@local-node"
-	// RemoteAuthServer is a special non-resolvable address that indicates client
-	// requests a connection to the remote auth server.
-	RemoteAuthServer = "@remote-auth-server"
-	// LocalKubernetes is a special non-resolvable address that indicates that clients
-	// requests a connection to the kubernetes endpoint of the local proxy.
-	// This has to be a valid domain name, so it lacks @
-	LocalKubernetes = "remote.kube.proxy." + constants.APIDomain
-	// LocalWindowsDesktop is a special non-resolvable address that indicates
-	// that clients requests a connection to the windows service endpoint of
-	// the local proxy.
-	// This has to be a valid domain name, so it lacks @
-	LocalWindowsDesktop = "remote.windows_desktop.proxy." + constants.APIDomain
-)
