+++ lib/reversetunnel/transport.go
@@ -18,8 +18,6 @@ package reversetunnel
 
 import (
 	"context"
-	"crypto/tls"
-	"crypto/x509"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -32,103 +30,15 @@ import (
 	"golang.org/x/crypto/ssh"
 
 	"github.com/gravitational/teleport"
-	"github.com/gravitational/teleport/api/client"
 	apidefaults "github.com/gravitational/teleport/api/defaults"
 	"github.com/gravitational/teleport/api/types"
 	"github.com/gravitational/teleport/api/utils/sshutils"
 	"github.com/gravitational/teleport/lib/auth"
 	"github.com/gravitational/teleport/lib/events"
 	"github.com/gravitational/teleport/lib/multiplexer"
-	alpncommon "github.com/gravitational/teleport/lib/srv/alpnproxy/common"
+	"github.com/gravitational/teleport/lib/reversetunnelclient"
 	"github.com/gravitational/teleport/lib/utils"
-	"github.com/gravitational/teleport/lib/utils/proxy"
 )
-
-// NewTunnelAuthDialer creates a new instance of TunnelAuthDialer
-func NewTunnelAuthDialer(config TunnelAuthDialerConfig) (*TunnelAuthDialer, error) {
-	if err := config.CheckAndSetDefaults(); err != nil {
-		return nil, trace.Wrap(err)
-	}
-	return &TunnelAuthDialer{
-		TunnelAuthDialerConfig: config,
-	}, nil
-}
-
-// TunnelAuthDialerConfig specifies TunnelAuthDialer configuration.
-type TunnelAuthDialerConfig struct {
-	// Resolver retrieves the address of the proxy
-	Resolver Resolver
-	// ClientConfig is SSH tunnel client config
-	ClientConfig *ssh.ClientConfig
-	// Log is used for logging.
-	Log logrus.FieldLogger
-	// InsecureSkipTLSVerify is whether to skip certificate validation.
-	InsecureSkipTLSVerify bool
-	// ClusterCAs contains cluster CAs.
-	ClusterCAs *x509.CertPool
-}
-
-func (c *TunnelAuthDialerConfig) CheckAndSetDefaults() error {
-	if c.Resolver == nil {
-		return trace.BadParameter("missing tunnel address resolver")
-	}
-	if c.ClusterCAs == nil {
-		return trace.BadParameter("missing cluster CAs")
-	}
-	return nil
-}
-
-// TunnelAuthDialer connects to the Auth Server through the reverse tunnel.
-type TunnelAuthDialer struct {
-	// TunnelAuthDialerConfig is the TunnelAuthDialer configuration.
-	TunnelAuthDialerConfig
-}
-
-// DialContext dials auth server via SSH tunnel
-func (t *TunnelAuthDialer) DialContext(ctx context.Context, _, _ string) (net.Conn, error) {
-	// Connect to the reverse tunnel server.
-	opts := []proxy.DialerOptionFunc{
-		proxy.WithInsecureSkipTLSVerify(t.InsecureSkipTLSVerify),
-	}
-
-	addr, mode, err := t.Resolver(ctx)
-	if err != nil {
-		t.Log.Errorf("Failed to resolve tunnel address: %v", err)
-		return nil, trace.Wrap(err)
-	}
-
-	if mode == types.ProxyListenerMode_Multiplex {
-		opts = append(opts, proxy.WithALPNDialer(client.ALPNDialerConfig{
-			TLSConfig: &tls.Config{
-				NextProtos:         []string{string(alpncommon.ProtocolReverseTunnel)},
-				InsecureSkipVerify: t.InsecureSkipTLSVerify,
-			},
-			DialTimeout:             t.ClientConfig.Timeout,
-			ALPNConnUpgradeRequired: client.IsALPNConnUpgradeRequired(addr.Addr, t.InsecureSkipTLSVerify),
-			GetClusterCAs:           client.ClusterCAsFromCertPool(t.ClusterCAs),
-		}))
-	}
-
-	dialer := proxy.DialerFromEnvironment(addr.Addr, opts...)
-	sconn, err := dialer.Dial(ctx, addr.AddrNetwork, addr.Addr, t.ClientConfig)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	// Build a net.Conn over the tunnel. Make this an exclusive connection:
-	// close the net.Conn as well as the channel upon close.
-	conn, _, err := sshutils.ConnectProxyTransport(
-		sconn.Conn,
-		&sshutils.DialReq{
-			Address: RemoteAuthServer,
-		},
-		true,
-	)
-	if err != nil {
-		return nil, trace.NewAggregate(err, sconn.Close())
-	}
-	return conn, nil
-}
 
 // parseDialReq parses the dial request. Is backward compatible with legacy
 // payload.
@@ -247,7 +157,7 @@ func (p *transport) start() {
 	// Handle special non-resolvable addresses first.
 	switch dreq.Address {
 	// Connect to an Auth Server.
-	case RemoteAuthServer:
+	case reversetunnelclient.RemoteAuthServer:
 		if len(p.authServers) == 0 {
 			p.log.Errorf("connection rejected: no auth servers configured")
 			p.reply(req, false, []byte("no auth servers configured"))
@@ -257,7 +167,7 @@ func (p *transport) start() {
 
 		directAddress = utils.ChooseRandomString(p.authServers)
 	// Connect to the Kubernetes proxy.
-	case LocalKubernetes:
+	case reversetunnelclient.LocalKubernetes:
 		switch p.component {
 		case teleport.ComponentReverseTunnelServer:
 			p.reply(req, false, []byte("connection rejected: no remote kubernetes proxy"))
@@ -300,7 +210,7 @@ func (p *transport) start() {
 		}
 
 	// LocalNode requests are for the single server running in the agent pool.
-	case LocalNode, LocalWindowsDesktop:
+	case reversetunnelclient.LocalNode, reversetunnelclient.LocalWindowsDesktop:
 		// Transport is allocated with both teleport.ComponentReverseTunnelAgent
 		// and teleport.ComponentReverseTunnelServer. However, dialing to this address
 		// only makes sense when running within a teleport.ComponentReverseTunnelAgent.
@@ -362,7 +272,7 @@ func (p *transport) start() {
 		clientDst = dst
 	}
 	var signedHeader []byte
-	isKubeOrAuth := dreq.ConnType == types.KubeTunnel || dreq.Address == RemoteAuthServer
+	isKubeOrAuth := dreq.ConnType == types.KubeTunnel || dreq.Address == reversetunnelclient.RemoteAuthServer
 	if shouldSendSignedPROXYHeader(p.proxySigner, dreq.TeleportVersion, useTunnel, !isKubeOrAuth, clientSrc, clientDst) {
 		signedHeader, err = p.proxySigner.SignPROXYHeader(clientSrc, clientDst)
 		if err != nil {
